+++ src/extension.js
@@ -1,8 +1,15 @@
 const vscode = require('vscode');
 const path = require("path");
+const fs = require("fs");
+const os = require("os");
+
+let myContext;
+let settings = vscode.workspace.getConfiguration("project-notes");
+let globalNotesFolder = settings.get("globalNotesFolder");
 
 // ========================================================================== //
-// ---=== Function Activate (Extension Activation) ===---
+//                     ---=== [Function Activate] ===---
+//                          (Extension Activation)
 // ========================================================================== //
 /**
  * @param {vscode.ExtensionContext} context
@@ -10,20 +17,35 @@ const path = require("path");
 function activate(context) {
 
     // ========================================================================== //
+    //      Activate - Initialize Extension
+    myContext = context;
+    
+    // ========================================================================== //
     //      Activate - Register Extension Commands
     vscode.commands.registerCommand('project-notes.openorcreatenote', openOrCreateNote);
-    vscode.commands.registerCommand('project-notes.browsenotesedit', browseNotesEdit);
-    vscode.commands.registerCommand('project-notes.browsenotespreview', browseNotesPreview);
+    vscode.commands.registerCommand('project-notes.notesedit', notesEdit);
+    vscode.commands.registerCommand('project-notes.notespreview', notesPreview);
+    vscode.commands.registerCommand('project-notes.setnotesglobalfolder', setNotesGlobalFolder);
+    vscode.commands.registerCommand('project-notes.openorcreateglobalnote', openOrCreateGlobalNote);
+    vscode.commands.registerCommand('project-notes.notesglobaledit', notesGlobalEdit);
+    vscode.commands.registerCommand('project-notes.notesglobalpreview', notesGlobalPreview);
 
     // ========================================================================== //
     //      Activate - Push Subscriptions
     context.subscriptions.push(openOrCreateNote);
-    context.subscriptions.push(browseNotesEdit);
-    context.subscriptions.push(browseNotesPreview);
+    context.subscriptions.push(notesEdit);
+    context.subscriptions.push(notesPreview);
+    context.subscriptions.push(setNotesGlobalFolder);
+    context.subscriptions.push(openOrCreateGlobalNote);
+    context.subscriptions.push(notesGlobalEdit);
+    context.subscriptions.push(notesGlobalPreview);
 
 }
+
+
 // ========================================================================== //
-// ---=== Function openOrCreateNote - (Open or Create New Note) ===---
+//                 ---=== [Function openOrCreateNote] ===---
+//                         (Open or Create New Note)
 // ========================================================================== //
 async function openOrCreateNote() {
 
@@ -43,7 +65,7 @@ async function openOrCreateNote() {
     // ========================================================================== //
     //      openOrCreateNote - Get current lines text
     const lineText = editor.document.lineAt(editor.selection.active.line).text;
-    const fileregex = new RegExp(/file: *([A-Za-z0-9_-]+)/);
+    const fileregex = new RegExp(/file: *([A-Za-z0-9_-]+)/i);
     var found = fileregex.test(lineText);
 
     // ========================================================================== //
@@ -82,12 +104,13 @@ async function openOrCreateNote() {
 
 
 // ========================================================================== //
-// ---=== Function browseNotesEdit - (Open Existing Project Note in Edit Mode) ===---
+//                     ---=== [Function notesEdit] ===---
+//                 (Open Existing Project Note in Edit Mode) 
 // ========================================================================== //
-async function browseNotesEdit() {
+async function notesEdit() {
 
     // ========================================================================== //
-    //      browseNotesEdit - Verify Workspace Folder Exists
+    //      notesEdit - Verify Workspace Folder Exists
     const workspaceFolders = vscode.workspace.workspaceFolders;
     if (!workspaceFolders || workspaceFolders.length === 0) {
         vscode.window.showWarningMessage('Workspace Folder Not Available!');
@@ -95,7 +118,7 @@ async function browseNotesEdit() {
     }
 
     // ========================================================================== //
-    //      browseNotesEdit - Initialize Quick Pick
+    //      notesEdit - Initialize Quick Pick
     let items = [];
     let options = {
         placeHolder: "Select Project Note to Edit",
@@ -103,7 +126,7 @@ async function browseNotesEdit() {
     };
 
     // ========================================================================== //
-    //      browseNotesEdit - Search for Project Notes and Show Quick Pick
+    //      notesEdit - Search for Project Notes and Show Quick Pick
     let inputAsWorkspaceRelativeFolder = '.vscode';
     let glob = '**/' + inputAsWorkspaceRelativeFolder + '/*.MD';
     let results = await vscode.workspace.findFiles(glob, null, 500);
@@ -118,7 +141,7 @@ async function browseNotesEdit() {
     const result = await vscode.window.showQuickPick(items, options);
 
     // ========================================================================== //
-    //      browseNotesEdit - Open Project Note File or Return if Cancelled
+    //      notesEdit - Open Project Note File or Return if Cancelled
     if (result == undefined) {
         // If Result = `undefined`, No File was Picked so Return
         return;
@@ -130,12 +153,13 @@ async function browseNotesEdit() {
 
 
 // ========================================================================== //
-// ---=== Function browseNotesPreview - (Open Existing Project Note in Preview Mode) ===---
+//                   ---=== [Function notesPreview]  ===---
+//                (Open Existing Project Note in Preview Mode)
 // ========================================================================== //
-async function browseNotesPreview() {
+async function notesPreview() {
 
     // ========================================================================== //
-    //      browseNotesPreview - Verify Workspace Folder Exists
+    //      notesPreview - Verify Workspace Folder Exists
     const workspaceFolders = vscode.workspace.workspaceFolders;
     if (!workspaceFolders || workspaceFolders.length === 0) {
         vscode.window.showWarningMessage('Workspace Folder Not Available!');
@@ -143,15 +167,15 @@ async function browseNotesPreview() {
     }
 
     // ========================================================================== //
-    //      browseNotesPreview - Initialize Quick Pick
+    //      notesPreview - Initialize Quick Pick
     let items = [];
     let options = {
-        placeHolder: "Select Project Note to Edit",
+        placeHolder: "Select Project Note to Preview",
         title: "---=== Project Notes - View Note in Preview Mode ===---"
     };
 
     // ========================================================================== //
-    //      browseNotesPreview - Search for Project Notes and Show Quick Pick
+    //      notesPreview - Search for Project Notes and Show Quick Pick
     let inputAsWorkspaceRelativeFolder = '.vscode';
     let glob = '**/' + inputAsWorkspaceRelativeFolder + '/*.MD';
     let results = await vscode.workspace.findFiles(glob, null, 500);
@@ -166,7 +190,7 @@ async function browseNotesPreview() {
     const result = await vscode.window.showQuickPick(items, options);
 
     // ========================================================================== //
-    //      browseNotesPreview - Open Project Note File or Return if Cancelled
+    //      notesPreview - Open Project Note File or Return if Cancelled
     if (result == undefined) {
         // If Result = `undefined`, No File was Picked so Return
         return;
@@ -175,14 +199,187 @@ async function browseNotesPreview() {
     await vscode.commands.executeCommand('markdown.showPreview', document);
 }
 
+// ========================================================================== //
+//               ---=== [Function setNotesGlobalFolder] ===---
+//                    (Set Global Notes Folder Location)
+// ========================================================================== //
+async function setNotesGlobalFolder() {
+    // Get Global Notes Folder From User
+    const home = vscode.Uri.file(path.join(os.homedir()))
+    const options = OpenDialogOptions = {
+        title: "Select Folder Location for Global Notes Storage",
+        defaultUri: home,
+        canSelectMany: false,
+        canSelectFolders: true,
+        canSelectFiles: false,
+        openLabel: "Select Folder for Global Notes Storage"
+    };
+    const folderUri = await vscode.window.showOpenDialog(options);
+    if (folderUri && folderUri[0]) {
+        globalNotesFolder = folderUri[0].fsPath;
+        //console.log('Selected file: ' + folderUri[0].fsPath);
+        console.log(globalNotesFolder);
+        let settings = vscode.workspace.getConfiguration("project-notes");
+        settings.update("globalNotesFolder",globalNotesFolder,1);
+    }
+}
+
+
 
 // ========================================================================== //
-// ---=== Function deactivate - This Method is Called When Extension is Deactivated ===---
+//                 ---=== [Function openOrCreateGlobalNote] ===---
+//                         (Create New Global Note)
+// ========================================================================== //
+async function openOrCreateGlobalNote() {
+    // ========================================================================== //
+    //      openOrCreateGlobalNote - Verify Text Editor Open + Global Notes Folder Exists
+    let editor = vscode.window.activeTextEditor;
+    if (!editor) {
+        vscode.window.showWarningMessage('Text Editor Not Open!');
+        return;
+    }
+    if (globalNotesFolder.length == 0) {
+        vscode.window.showWarningMessage('Global Folder Location has not been defined yet.');
+        return;
+
+    }
+
+    // ========================================================================== //
+    //      openOrCreateGlobalNote - Get current lines text
+    const lineText = editor.document.lineAt(editor.selection.active.line).text;
+    const globalfileregex = new RegExp(/global file: *([A-Za-z0-9_-]+)/i);
+    var found = globalfileregex.test(lineText);
+    let fileName;
+    let filePath;
+
+    // ========================================================================== //
+    //      openOrCreateGlobalNote - Get Filename from comment
+    if (found) {
+        const filenameArray = globalfileregex.exec(lineText);
+        fileName = filenameArray[1] + '.MD';
+        filePath = globalNotesFolder+'\\'+fileName;
+        // ========================================================================== //
+        //      openOrCreateGlobalNote - Prompt for Filename from User
+    } else {
+        fileName = await vscode.window.showInputBox({
+        placeHolder: "Global Note Name to Open or Create",
+        prompt: "Open or Create New Global Note: "
+        });
+        let parts = fileName.split(".");
+        fileName = parts[0]+'.MD';
+        filePath = globalNotesFolder+'\\'+fileName;
+    }
+
+    // ========================================================================== //
+    //      openOrCreateGlobalNote - Open Filename.MD
+    const activeTextEditor = vscode.window.activeTextEditor;
+    if (!activeTextEditor || activeTextEditor.document.fileName !== filePath) {
+        const workspaceEdit = new vscode.WorkspaceEdit();
+        workspaceEdit.createFile(vscode.Uri.file(filePath), { overwrite: false, ignoreIfExists: true });
+        await vscode.workspace.applyEdit(workspaceEdit);
+        const document = await vscode.workspace.openTextDocument(filePath);
+        vscode.window.showTextDocument(document, { preview: true });
+    } else {
+        if (activeTextEditor.document.isDirty) { // Save modified file first
+            const saved = await activeTextEditor.document.save();
+            if (saved) {
+                vscode.commands.executeCommand("workbench.action.closeActiveEditor");
+            }
+        } else {
+            vscode.commands.executeCommand("workbench.action.closeActiveEditor");
+        }
+    }
+}
+
+
+// ========================================================================== //
+//                  ---=== [Function notesGlobalEdit] ===---
+//                  (Open Existing Global Note in Edit Mode)
+// ========================================================================== //
+async function notesGlobalEdit() {
+    if (globalNotesFolder.length == 0) {
+        vscode.window.showWarningMessage('Global Folder Location has not been defined yet.');
+        return;
+    }
+    let items = [];
+    let options = {
+        placeHolder: "Select Global Note to Preview",
+        title: "---=== Project Notes - View Global Note in Preview Mode ===---"
+    };
+
+    // ========================================================================== //
+    //      notesGlobalPreview - Search for Global Notes and Show Quick Pick
+    results = await vscode.workspace.findFiles(new vscode.RelativePattern(globalNotesFolder, '*.md'), '**/node_modules/**');
+    if (results.length == 0) { // If Zero Note Files Found, Inform User and Return
+        vscode.window.showWarningMessage('No Global Note Files Found!');
+        return;
+    }
+    for (const file of results) {
+        const base = path.basename(file.fsPath);
+        items.push(base);
+    }
+    const result = await vscode.window.showQuickPick(items, options);
+
+    // ========================================================================== //
+    //      notesGlobalPreview - Open Global Note File or Return if Cancelled
+    if (result == undefined) {
+        // If Result = `undefined`, No File was Picked so Return
+        return;
+    }
+    const notesFilePath = path.join(globalNotesFolder, '\\') + result;
+    document = vscode.workspace.openTextDocument(notesFilePath);
+    await vscode.window.showTextDocument(document);
+}
+
+
+// ========================================================================== //
+//                ---=== [Function notesGlobalPreview] ===---
+//                (Open Existing Global Note in Preview Mode)
+// ========================================================================== //
+async function notesGlobalPreview() {
+    if (globalNotesFolder.length == 0) {
+        vscode.window.showWarningMessage('Global Folder Location has not been defined yet.');
+        return;
+    }
+    let items = [];
+    let options = {
+        placeHolder: "Select Global Note to Preview",
+        title: "---=== Project Notes - View Global Note in Preview Mode ===---"
+    };
+
+    // ========================================================================== //
+    //      notesGlobalPreview - Search for Global Notes and Show Quick Pick
+    results = await vscode.workspace.findFiles(new vscode.RelativePattern(globalNotesFolder, '*.md'), '**/node_modules/**');
+    if (results.length == 0) { // If Zero Note Files Found, Inform User and Return
+        vscode.window.showWarningMessage('No Global Note Files Found!');
+        return;
+    }
+    for (const file of results) {
+        const base = path.basename(file.fsPath);
+        items.push(base);
+    }
+    const result = await vscode.window.showQuickPick(items, options);
+
+    // ========================================================================== //
+    //      notesGlobalPreview - Open Global Note File or Return if Cancelled
+    if (result == undefined) {
+        // If Result = `undefined`, No File was Picked so Return
+        return;
+    }
+    const document = vscode.Uri.file(path.join(globalNotesFolder,'/') + result);
+    await vscode.commands.executeCommand('markdown.showPreview', document);
+}
+
+
+// ========================================================================== //
+//                 ---=== [Function deactivate] ===---
+//         (This Method is Called When Extension is Deactivated)
 // ========================================================================== //
 function deactivate() {}
 
+
 // ========================================================================== //
-// ---=== Export modules ===---
+//                       ---=== [Export modules] ===---
 // ========================================================================== //
 module.exports = {
     activate,